Adds CAN FD flag on message object.
Adds CAN FD flag on message object when generating CPP file for the
low-can service. So with this commit you are now able to use a new flags
in your JSON CAN message definitions. This will imply that messages
received from a certain CAN ID will be handled as CAN FD.

Bug-AGL: SPEC-1980

Change-Id: Ia5fb573711742591c068928aee914ba708c802df
Signed-off-by: Romain Forlot <romain.forlot@iot.bzh>